backend/auth_service/app/__init__.py
# 主程序
from flask import Flask
from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate
from flask_bcrypt import Bcrypt
from flask_cors import CORS
from config import Config
import jwt
import datetime

db = SQLAlchemy()
migrate = Migrate()
bcrypt = Bcrypt()

def create_app():
    app = Flask(__name__)
    app.config.from_object(Config)

    db.init_app(app)
    migrate.init_app(app, db)
    bcrypt.init_app(app)
    CORS(app) # Enable CORS for frontend

    from app import routes, models # Import routes and models here
    app.register_blueprint(routes.auth_bp)

    return app